+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,17 +1,15 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.techcourse.config.DataSourceConfig;
+import com.interface21.transaction.TransactionManager;
 import com.techcourse.domain.User;
-import java.sql.Connection;
-import java.sql.SQLException;
 
 public class TxUserService implements UserService {
 
+    private final TransactionManager transactionManager;
     private final UserService userService;
 
-    public TxUserService(final UserService userService) {
+    public TxUserService(final TransactionManager transactionManager, final UserService userService) {
+        this.transactionManager = transactionManager;
         this.userService = userService;
     }
 
@@ -27,25 +25,6 @@ public class TxUserService implements UserService {
 
     @Override
     public void changePassword(final long id, final String newPassword, final String createdBy) {
-        final var dataSource = DataSourceConfig.getInstance();
-        final var conn = DataSourceUtils.getConnection(dataSource);
-        try {
-            conn.setAutoCommit(false);
-            userService.changePassword(id, newPassword, createdBy);
-            conn.commit();
-        } catch (SQLException e) {
-            rollback(conn);
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(conn, dataSource);
-        }
-    }
-
-    private void rollback(final Connection conn) {
-        try {
-            conn.rollback();
-        } catch (SQLException e) {
-            throw new DataAccessException(e);
-        }
+        transactionManager.doInTransaction(conn -> userService.changePassword(id, newPassword, createdBy));
     }
 }
