+++ jdbc/src/main/java/com/interface21/transaction/TransactionManager.java
@@ -1,0 +1,39 @@
+package com.interface21.transaction;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.function.Consumer;
+import javax.sql.DataSource;
+
+public class TransactionManager {
+
+    private final DataSource dataSource;
+
+    public TransactionManager(final DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public void doInTransaction(final Consumer<Connection> consumer) {
+        final var conn = DataSourceUtils.getConnection(dataSource);
+        try {
+            conn.setAutoCommit(false);
+            consumer.accept(conn);
+            conn.commit();
+        } catch (SQLException | DataAccessException e) {
+            rollback(conn);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(conn, dataSource);
+        }
+    }
+
+    private void rollback(final Connection conn) {
+        try {
+            conn.rollback();
+        } catch (SQLException e) {
+            throw new DataAccessException(e);
+        }
+    }
+}
